refactor(review): split training/{config,utils}.py into focused modules (Theme F)
config.py and utils.py had grown to 1.3k and 1.5k LOC, mixing archive
fingerprinting, patch extraction, metric trackers, baseline IO, and the
core TissueNetConfig/RNG/log helpers in one place each. Carve four
focused modules out (verbatim, no logic changes):

- training/archive.py: zarr v3 alpha metadata patch, archive metadata
  / array fingerprinting, FOV-key discovery, and the per-process caches.
- training/patch.py: per-cell patch extraction
  (compute_distance_transform, extract_patch_from_zarr, extract_patch).
- training/metrics.py: confusion-matrix hierarchy adjustment,
  MP per-marker reduction, MPMetricsTracker, LossesAndMetrics,
  build_label_remap.
- training/baseline_features.py: baseline classifier feature extraction
  pipeline (_conf_mat_summary, compute_baseline_metrics,
  save_baseline_predictions, _extract_all_dataset_features,
  extract_features_from_zarr, _get_cell_data_from_ds).

Re-exports at the bottom of config.py and utils.py keep all
tests/scripts working unchanged (230 passed, 1 skipped, matching the
pre-split baseline). dataset.py is updated to import directly from
the new homes for cached_archive_metadata_fingerprint and extract_patch.

Two non-mechanical touches required to keep monkey-patch-based tests
green:
- baseline_features.extract_features_from_zarr looks up
  _discover_fov_keys and _extract_all_dataset_features via the
  config / utils modules at call time, so tests that monkeypatch
  those symbols on the legacy modules still take effect after the
  split. _FINGERPRINT_CACHE / _FOV_KEYS_CACHE dicts are re-exported
  from config.py for the same reason (test_dataset_cache mutates them).
- metrics.LossesAndMetrics.compute defers import of _conf_mat_summary
  to method-call time to avoid a metrics <-> baseline_features import
  cycle (baseline_features needs adjust_conf_mat_hierarchy at module
  load).
